New features
- Remove KMSectionHeaderView ’s UIImagView property
- Change headerHeight to CGFloat
- Rename KMSection properties
- New DataSource methods for Animations (close and open)
- New reload section methods
- example code improved

+++ Classes/Model/KMAppearence.h
@@ -6,7 +6,7 @@
 
 @interface KMAppearence : NSObject
 
-@property(nonatomic, assign) NSInteger headerHeight;
+@property(nonatomic, assign) CGFloat headerHeight;
 @property(nonatomic, strong) UIFont *headerFont;
 @property(nonatomic, strong) UIColor *headerTitleColor;
 @property(nonatomic, strong) UIColor *headerColor;
